// src/enums/permissionCategory.ts
export enum PermissionCategory {
  DEVICE_MANAGEMENT = 'Device Management',
  DEVICE_USER = 'Device User',
  NOTIFICATION = 'Notification',
  SERVICE_ACCOUNT = 'Service Account',
  PRICE_MANAGEMENT = 'Price Management',
  FINANCIAL = 'Financial',
  USAGE_STATS = 'Usage Stats',
  PAYMENT = 'Payment',
}